# docs/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Betsi Patient Flow API",
    "description": "Patient flow and escalation for emergency departments. Errors are RFC 9457 problem details with a stable \u0027code\u0027. See docs/API.md and docs/API-VERSIONING.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/escalation-policy": {
      "get": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "The revision in force, and every revision with its approval history.",
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOverview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/revisions/{revision}": {
      "get": {
        "tags": [
          "EscalationPolicy"
        ],
        "parameters": [
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyRevisionView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/preview": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "Shows how many patients waiting right now each proposed tier would apply to. Changes nothing.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WaitingTimeTierInput"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WaitingTimeTierInput"
                }
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WaitingTimeTierInput"
                }
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyPreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/proposals": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "Proposes a new revision. It has no effect until approved.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyCommand"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyCommand"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/proposals/restore": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "Proposes a new revision restoring an earlier approved revision.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyRestorationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyRestorationCommand"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ProposeEscalationPolicyRestorationCommand"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/{policyId}/approve": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "Approves a proposal. Must be a different, supervisory user from the proposer.",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveEscalationPolicyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveEscalationPolicyCommand"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ApproveEscalationPolicyCommand"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/{policyId}/reject": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectEscalationPolicyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectEscalationPolicyCommand"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RejectEscalationPolicyCommand"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalation-policy/{policyId}/withdraw": {
      "post": {
        "tags": [
          "EscalationPolicy"
        ],
        "summary": "Withdraws a proposal, or an approval that has not yet taken effect.",
        "parameters": [
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawEscalationPolicyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawEscalationPolicyCommand"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawEscalationPolicyCommand"
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveEscalationPolicyCommand": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "effectiveFrom": {
            "type": "string",
            "description": "UTC. Omit to take effect immediately.",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CommandResult": {
        "type": "object",
        "properties": {
          "aggregateId": {
            "type": "string",
            "description": "Identifier of the aggregate the command created or changed.",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "description": "The aggregate\u0027s version after the command was applied.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "What a command produced: the aggregate it affected and the version it left behind."
      },
...
      "PolicyOverview": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "inForce": {
            "$ref": "#/components/schemas/PolicyRevisionView"
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyRevisionView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicyPreview": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "patientsWaiting": {
            "type": "integer",
            "format": "int32"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyPreviewTier"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "What a candidate set of tiers would do to the patients waiting right now."
      },
      "PolicyPreviewTier": {
        "type": "object",
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "thresholdMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "responsibleRole": {
            "type": "string",
            "nullable": true
          },
          "patientsAtOrPastThreshold": {
            "type": "integer",
            "format": "int32"
          },
          "wouldBeNewlyEscalated": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PolicyRevisionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "inForce": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyTierView"
            },
            "nullable": true
          },
          "followUpOwnerRole": {
            "type": "string",
            "nullable": true
          },
          "restoresRevision": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "proposalReason": {
            "type": "string",
            "nullable": true
          },
          "proposedByActorId": {
            "type": "string",
            "format": "uuid"
          },
          "proposedByRole": {
            "type": "string",
            "nullable": true
          },
          "proposedAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedByActorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "decidedByRole": {
            "type": "string",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "decisionReason": {
            "type": "string",
            "nullable": true
          },
          "effectiveFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PolicyTierView": {
        "type": "object",
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "thresholdMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "responsibleRole": {
            "type": "string",
            "nullable": true
          },
          "acknowledgementDeadlineMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "recommendedAction": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "ProposeEscalationPolicyCommand": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "tiers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WaitingTimeTierInput"
            },
            "description": "In threshold order. Levels are assigned 1, 2, 3\u2026 in the order given.",
            "nullable": true
          },
          "followUpOwnerRole": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Proposes a new policy revision. Takes no effect until someone else approves it."
      },
      "ProposeEscalationPolicyRestorationCommand": {
        "type": "object",
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Proposes a new revision restoring an earlier approved one (rollback)."
      },
...
      "RejectEscalationPolicyCommand": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WaitingTimeTierInput": {
        "type": "object",
        "properties": {
          "thresholdMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "responsibleRole": {
            "type": "string",
            "nullable": true
          },
          "acknowledgementDeadlineMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "recommendedAction": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WithdrawEscalationPolicyCommand": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "X-Betsi-Tenant": {
        "type": "apiKey",
        "description": "Tenant ID. Development tenants: 11111111-1111-1111-1111-111111111111 or 22222222-2222-2222-2222-222222222222.",
        "name": "X-Betsi-Tenant",
        "in": "header"
      },
      "X-Betsi-Actor-Role": {
        "type": "apiKey",
        "description": "Actor role, e.g. Nurse.",
        "name": "X-Betsi-Actor-Role",
        "in": "header"
      },
      "X-Betsi-Actor": {
        "type": "apiKey",
        "description": "Optional actor ID (GUID).",
        "name": "X-Betsi-Actor",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "An access token from the configured OIDC provider. In Development the X-Betsi headers may be used instead.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [],
      "X-Betsi-Tenant": [],
      "X-Betsi-Actor-Role": [],
      "X-Betsi-Actor": []
    }
  ],
...
}